Stop catching errors: crash early
Catching errors will allow broken deploys to continue rather silently.
This update removes the catching of errors and instead adds the
_throwing_ of an error when the .env file is not found.
There's no good reason to continue the deploy when you know the
environment variables can't be loaded.

If the file is not there and shouldn't be there, there's no reason to
use this plugin.